Added UI and lot of other features

+++ lightshow/audio/audio_streams.py
@@ -1,19 +1,22 @@
 from collections import deque
+import traceback
 from typing import List, Type, Callable, Optional
 
 import numpy as np
 import pyaudiowpatch as pyaudio
 
 from lightshow.audio.audio_types import (
-    AudioData,
+    AAudioCapture,
+    AAudioStreamHandler,
     AudioListener,
     Processor,
     AudioListenerType,
 )
+from lightshow.gui.utils.message_box import post_ui_message
 from lightshow.utils.config import Config
 
 
-class AudioCapture:
+class AudioCapture(AAudioCapture):
     """
     Collects audio data via a PyAudio callback.
     """
@@ -21,11 +24,13 @@ class AudioCapture:
     def __init__(
         self,
         processor: Processor,
+        stream_handler: AAudioStreamHandler,
         chunk_size=512,
         max_buffer=10,
         channels=1,
         sample_rate=44100,
     ):
+        self.stream_handler = stream_handler
         self.sample_rate = sample_rate
         self.chunk_size = chunk_size
         self.audio_buffer = deque(maxlen=max_buffer)
@@ -57,7 +62,7 @@ class AudioCapture:
         """
 
         if isinstance(listener, type) and issubclass(listener, AudioListener):
-            listener = listener(self.channels, self.chunk_size, self.sample_rate)
+            listener = listener(self.stream_handler)
             self.listeners.append(listener)
             return listener
         elif isinstance(listener, AudioListener):
@@ -73,35 +78,66 @@ class AudioCapture:
     def remove_listener(self, listener: Type[AudioListener]):
         if not issubclass(listener, AudioListener):
             raise ValueError("Listener must be an instance of AudioListener")
-        self.listeners.remove(
-            listener(self.channels, self.chunk_size, self.sample_rate)
-        )
+        self.listeners.remove(listener(self.stream_handler))
 
     def get_latest_data(self):
         return self.audio_buffer[-1] if self.audio_buffer else None
 
 
-class AudioStreamHandler:
+class AudioStreamHandler(AAudioStreamHandler):
     """
     Sets up the audio device (using WASAPI loopback when available) and manages the stream.
     """
 
     def __init__(self, processor: Type[Processor], config: Config):
-        self.chunk_size = config.chunk_size
+        self.chunk_size = config.chunk_size or 1024
         self.pyaudio_instance = pyaudio.PyAudio()
-        self.config = config
-        self.stream = None
-        self.device_index = None
-        self.setup_device()
-        self.processor = processor(self.chunk_size, self.sample_rate)
-        self.audio_capture = AudioCapture(
-            self.processor,
-            chunk_size=self.chunk_size,
-            channels=self.channels,
-            sample_rate=self.sample_rate,
-        )
+        self.processor_class = processor
+        self.device_change_listeners: List[Callable[[AudioCapture], None]] = []
+        self.pending_listeners = []
+        # self.reinit_stream(config)
+
+    def add_device_change_listener(self, listener: Callable[[AudioCapture], None]):
+        self.device_change_listeners.append(listener)
+
+    def reinit_stream(self, config: Config):
+        """Initializes or re-initializes the stream with a new configuration."""
+        try:
+            if hasattr(self, "stream") and self.stream and self.stream.is_active():
+                self.stop_stream()
+
+            self.config = config
+            self.stream = None
+            self.device_index = None
+            self.setup_device()
+            self.processor = self.processor_class(self.chunk_size, self.sample_rate)
+
+            # Preserve listeners if AudioCapture already exists
+            old_listeners = []
+            if hasattr(self, "audio_capture"):
+                old_listeners = self.audio_capture.listeners
+
+            self.audio_capture = AudioCapture(
+                self.processor,
+                self,
+                chunk_size=self.chunk_size,
+                channels=self.channels,
+                sample_rate=self.sample_rate,
+            )
+            for listener in self.device_change_listeners:
+                listener(self.audio_capture)
+            self.audio_capture.listeners = old_listeners
+            for listener in self.pending_listeners:
+                print("Adding listener")
+                self.audio_capture.add_listener(listener)
+            self.pending_listeners = []
+        except Exception:
+            print("init error")
+            post_ui_message("Audio Initialization Error", traceback.format_exc())
 
     def setup_device(self):
+        if not self.pyaudio_instance:
+            return
         try:
             wasapi_info = self.pyaudio_instance.get_host_api_info_by_type(
                 pyaudio.paWASAPI
@@ -141,20 +177,39 @@ class AudioStreamHandler:
             f"Using device: {default_device['name']} (index: {self.device_index}) Sample Rate: {self.sample_rate} Hz Channels: {self.channels} Chunk size: {self.chunk_size}"
         )
 
+    def add_listener_on_init(
+        self, listener: AudioListenerType
+    ) -> Optional[AudioListener]:
+        return self.pending_listeners.append(listener)
+
     def start_stream(self):
-        self.stream = self.pyaudio_instance.open(
-            format=pyaudio.paInt16,
-            channels=self.channels,
-            rate=self.sample_rate,
-            frames_per_buffer=self.chunk_size,
-            input=True,
-            input_device_index=self.device_index,
-            stream_callback=self.audio_capture.callback,
-        )
-        self.stream.start_stream()
+        try:
+            if not self.pyaudio_instance:
+                return
+            self.stream = self.pyaudio_instance.open(
+                format=pyaudio.paInt16,
+                channels=self.channels,
+                rate=self.sample_rate,
+                frames_per_buffer=self.chunk_size,
+                input=True,
+                input_device_index=self.device_index,
+                stream_callback=self.audio_capture.callback,
+            )
+            self.stream.start_stream()
+        except Exception:
+            print("Failed to start audio stream")
+            post_ui_message("Audio Stream Error", traceback.format_exc())
 
     def stop_stream(self):
-        if self.stream is not None:
-            self.stream.stop_stream()
+        if hasattr(self, "stream") and self.stream is not None:
+            if self.stream.is_active():
+                self.stream.stop_stream()
             self.stream.close()
-        self.pyaudio_instance.terminate()
+            self.stream = None
+
+    def close(self):
+        """Stops the stream and terminates the PyAudio instance."""
+        self.stop_stream()
+        if self.pyaudio_instance:
+            self.pyaudio_instance.terminate()
+            self.pyaudio_instance = None
